refactor pretty-print structure
